requirements.txt
transformers
torch
librosa
torchaudio
unidic-lite
mecab-python3
streamlit
audio-recorder-streamlit
st-annotated-text